[refactor] Put things in order
