+++ scripts/filter_locations.py
