+++ scripts/merge_sarif.py
